feat: migrate product storage from PostgreSQL to MongoDB

+++ ECommerce.Application/DTO/Request/Product/ProductUpdateRequestDto.cs
@@ -4,8 +4,8 @@ public record ProductUpdateRequestDto
 {
     public required string Name {get;set;}
     public required string Description {get;set;}
-    public required decimal Price {get;set;}
-    public required decimal DiscountRate {get;set;}
+    public required double Price {get;set;}
+    public required double DiscountRate {get;set;}
     public required string? ImageUrl {get;set;}
     public required int StockQuantity {get;set;}
     public required int CategoryId {get;set;}
