+++ ECommerce.Application/DTO/Response/BasketItem/BasketItemResponseDto.cs
@@ -4,7 +4,7 @@ public record BasketItemResponseDto
 {
     public int AccountId { get; set; }
     public int Quantity { get; init; }
-    public decimal UnitPrice { get; init; }
+    public double UnitPrice { get; init; }
     public int ProductId { get; set; }
     public required string ProductName { get; set; }
 }
