+++ ECommerce.Application/Events/ProductCreatedEvent.cs
@@ -7,8 +7,8 @@ public class ProductCreatedEvent : INotification
     public int ProductId { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
-    public decimal Price { get; set; }
-    public decimal DiscountRate { get; set; }
+    public double Price { get; set; }
+    public double DiscountRate { get; set; }
     public string ImageUrl { get; set; }
     public int StockQuantity { get; set; }
     public int CategoryId { get; set; }
